test: move EnsureRequirements tests to core/tests/physical_optimizer/
The ~1540-line test module that lived inside
`datafusion/physical-optimizer/src/ensure_requirements/mod.rs` is
now a sibling integration test under
`datafusion/core/tests/physical_optimizer/`, where the rest of the
physical-optimizer test suite lives.

This is an as-is move: every test still uses the
`MockMultiPartitionExec` / `MockReqExec` shims from the original module
and the same assertions. Subsequent commits remove redundant tests and
fold overlapping ones — keeping those changes separate makes the diff
easy to review.
